+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/anccare/details/AncPatientItemMapper.kt
@@ -18,7 +18,7 @@ package org.smartregister.fhircore.anc.ui.anccare.details
 
 import com.google.android.fhir.logicalId
 import org.smartregister.fhircore.anc.AncApplication
-import org.smartregister.fhircore.anc.data.model.AncPatientItem
+import org.smartregister.fhircore.anc.data.anc.model.AncPatientItem
 import org.smartregister.fhircore.anc.ui.anccare.register.Anc
 import org.smartregister.fhircore.engine.data.domain.util.DomainMapper
 import org.smartregister.fhircore.engine.util.extension.extractAge
