+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/family/register/FamilyRegisterActivity.kt
@@ -24,7 +24,9 @@ import androidx.fragment.app.Fragment
 import kotlinx.coroutines.runBlocking
 import org.smartregister.fhircore.anc.AncApplication
 import org.smartregister.fhircore.anc.R
-import org.smartregister.fhircore.anc.data.FamilyRepository
+import org.smartregister.fhircore.anc.data.anc.AncPatientRepository
+import org.smartregister.fhircore.anc.data.family.FamilyRepository
+import org.smartregister.fhircore.anc.ui.anccare.register.AncItemMapper
 import org.smartregister.fhircore.anc.ui.anccare.register.AncRegisterActivity
 import org.smartregister.fhircore.anc.ui.family.form.FamilyFormConstants
 import org.smartregister.fhircore.anc.ui.family.form.FamilyQuestionnaireActivity
@@ -35,6 +37,7 @@ import org.smartregister.fhircore.engine.ui.register.model.SideMenuOption
 
 class FamilyRegisterActivity : BaseRegisterActivity() {
   private lateinit var familyRepository: FamilyRepository
+  private lateinit var ancPatientRepository: AncPatientRepository
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
@@ -46,6 +49,9 @@ class FamilyRegisterActivity : BaseRegisterActivity() {
 
     familyRepository =
       FamilyRepository((application as AncApplication).fhirEngine, FamilyItemMapper)
+
+    ancPatientRepository =
+      AncPatientRepository((application as AncApplication).fhirEngine, AncItemMapper)
   }
 
   override fun sideMenuOptions(): List<SideMenuOption> =
@@ -62,7 +68,7 @@ class FamilyRegisterActivity : BaseRegisterActivity() {
         titleResource = R.string.anc_register_title,
         iconResource = ContextCompat.getDrawable(this, R.drawable.ic_baby_mother)!!,
         opensMainRegister = false,
-        countMethod = { runBlocking { familyRepository.ancPatientRepository.countAll() } }
+        countMethod = { runBlocking { ancPatientRepository.countAll() } }
       )
     )
