+++ android/anc/src/test/java/org/smartregister/fhircore/anc/AncFhirSyncWorkerTest.kt
@@ -52,8 +52,6 @@ class AncFhirSyncWorkerTest : RobolectricTest() {
     val data = ancFhirSyncWorker.getSyncData()
     Assert.assertEquals(3, data.size)
     Assert.assertTrue(data.containsKey(ResourceType.Patient))
-    Assert.assertTrue(data.containsKey(ResourceType.StructureMap))
-    Assert.assertTrue(data.containsKey(ResourceType.RelatedPerson))
   }
 
   @Test
